+++ src/bot.rs
@@ -2,26 +2,26 @@ use ahash::{HashMapExt as _, RandomState};
 use dialoguer::theme::ColorfulTheme;
 use dialoguer::{Input, Select};
 use plugin_builder::Listen;
+#[cfg(feature = "plugin-access-control")]
+use runtimebot::kovi_api::AccessList;
 use serde::{Deserialize, Serialize};
 use serde_json::{self, Value};
 use std::collections::{HashMap, HashSet};
 use std::env;
 use std::fmt::{Debug, Display};
-use std::future::Future;
 use std::io::Write as _;
 use std::net::{Ipv4Addr, Ipv6Addr};
-use std::pin::Pin;
 use std::{fs, net::IpAddr, sync::Arc};
 use tokio::sync::mpsc::{self};
-use tokio::sync::{oneshot, watch};
-use tokio::task::JoinHandle;
+use tokio::sync::watch;
 
 use crate::error::{BotBuildError, BotError};
-use crate::task::TASK_MANAGER;
 
+use crate::RT;
 #[cfg(feature = "plugin-access-control")]
 pub use crate::bot::runtimebot::kovi_api::AccessControlMode;
-use crate::RT;
+use crate::plugin::{Plugin, PluginStatus};
+use crate::types::KoviAsyncFn;
 
 pub(crate) mod connect;
 pub(crate) mod handler;
@@ -31,239 +31,18 @@ pub mod message;
 pub mod plugin_builder;
 pub mod runtimebot;
 
-tokio::task_local! {
-    pub static PLUGIN_BUILDER: crate::PluginBuilder;
+/// bot结构体
+#[derive(Clone)]
+pub struct Bot {
+    pub information: BotInformation,
+    pub(crate) plugins: HashMap<String, Plugin, RandomState>,
+    pub(crate) run_abort: Vec<tokio::task::AbortHandle>,
 }
-
-tokio::task_local! {
-    pub static PLUGIN_NAME: Arc<String>;
-}
-
-/// kovi的配置
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct KoviConf {
-    pub config: Config,
-    pub server: Server,
-}
-
-impl AsRef<KoviConf> for KoviConf {
-    fn as_ref(&self) -> &KoviConf {
-        self
-    }
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct Config {
-    pub main_admin: i64,
-    pub admins: Vec<i64>,
-    pub debug: bool,
-}
-
-impl KoviConf {
-    pub fn new(main_admin: i64, admins: Option<Vec<i64>>, server: Server, debug: bool) -> Self {
-        KoviConf {
-            config: Config {
-                main_admin,
-                admins: admins.unwrap_or_default(),
-                debug,
-            },
-            server,
-        }
-    }
-}
-
-type KoviAsyncFn = dyn Fn() -> Pin<Box<dyn Future<Output = ()> + Send>> + Send + Sync;
-
 impl Drop for Bot {
     fn drop(&mut self) {
         for i in self.run_abort.iter() {
             i.abort();
         }
-    }
-}
-
-/// bot结构体
-#[derive(Clone)]
-pub struct Bot {
-    pub information: BotInformation,
-    pub(crate) plugins: HashMap<String, BotPlugin, RandomState>,
-    pub(crate) run_abort: Vec<tokio::task::AbortHandle>,
-}
-
-#[derive(Clone)]
-pub(crate) struct BotPlugin {
-    pub(crate) enable_on_startup: bool,
-    pub(crate) enabled: watch::Sender<bool>,
-
-    pub(crate) name: String,
-    pub(crate) version: String,
-    pub(crate) main: Arc<KoviAsyncFn>,
-    pub(crate) listen: Listen,
-
-    #[cfg(feature = "plugin-access-control")]
-    pub(crate) access_control: bool,
-    #[cfg(feature = "plugin-access-control")]
-    pub(crate) list_mode: AccessControlMode,
-    #[cfg(feature = "plugin-access-control")]
-    pub(crate) access_list: AccessList,
-}
-
-#[cfg(feature = "plugin-access-control")]
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-pub struct AccessList {
-    pub friends: HashSet<i64>,
-    pub groups: HashSet<i64>,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct PluginInfo {
-    pub name: String,
-    pub version: String,
-    /// 插件是否启用
-    pub enabled: bool,
-    /// 插件是否在Bot启动时启用
-    pub enable_on_startup: bool,
-    /// 插件是否启用框架级访问控制
-    #[cfg(feature = "plugin-access-control")]
-    pub access_control: bool,
-    /// 插件的访问控制模式
-    #[cfg(feature = "plugin-access-control")]
-    pub list_mode: AccessControlMode,
-    /// 插件的访问控制列表
-    #[cfg(feature = "plugin-access-control")]
-    pub access_list: AccessList,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-struct PluginStatus {
-    enable_on_startup: bool,
-    #[cfg(feature = "plugin-access-control")]
-    access_control: bool,
-    #[cfg(feature = "plugin-access-control")]
-    list_mode: AccessControlMode,
-    #[cfg(feature = "plugin-access-control")]
-    access_list: AccessList,
-}
-
-/// bot信息结构体
-#[derive(Debug, Clone)]
-pub struct BotInformation {
-    pub main_admin: i64,
-    pub deputy_admins: HashSet<i64>,
-    pub server: Server,
-}
-/// server信息
-#[derive(Deserialize, Serialize, Debug, Clone)]
-pub struct Server {
-    pub host: Host,
-    pub port: u16,
-    pub access_token: String,
-    pub secure: bool,
-}
-
-#[derive(Deserialize, Serialize, Debug, Clone)]
-#[serde(untagged)]
-pub enum Host {
-    IpAddr(IpAddr),
-    Domain(String),
-}
-
-impl Display for Host {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Host::IpAddr(ip) => write!(f, "{}", ip),
-            Host::Domain(domain) => write!(f, "{}", domain),
-        }
-    }
-}
-
-impl Server {
-    pub fn new(host: Host, port: u16, access_token: String, secure: bool) -> Self {
-        Server {
-            host,
-            port,
-            access_token,
-            secure,
-        }
-    }
-}
-
-#[derive(Debug, Deserialize, Serialize, Clone)]
-pub struct SendApi {
-    pub action: String,
-    pub params: Value,
-    pub echo: String,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct ApiReturn {
-    pub status: String,
-    pub retcode: i32,
-    pub data: Value,
-    pub echo: String,
-}
-
-pub(crate) type ApiAndOneshot = (
-    SendApi,
-    Option<oneshot::Sender<Result<ApiReturn, ApiReturn>>>,
-);
-
-impl std::fmt::Display for ApiReturn {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "status: {}, retcode: {}, data: {}, echo: {}",
-            self.status, self.retcode, self.data, self.echo
-        )
-    }
-}
-
-impl std::fmt::Display for SendApi {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", serde_json::to_string(self).unwrap())
-    }
-}
-
-impl SendApi {
-    pub fn new(action: &str, params: Value, echo: &str) -> Self {
-        SendApi {
-            action: action.to_string(),
-            params,
-            echo: echo.to_string(),
-        }
-    }
-}
-
-impl BotPlugin {
-    fn shutdown(&mut self) -> JoinHandle<()> {
-        log::debug!("Plugin '{}' is dropping.", self.name,);
-
-        let plugin_name_ = Arc::new(self.name.clone());
-
-        let mut task_vec = Vec::new();
-
-        for listen in &self.listen.drop {
-            let listen_clone = listen.clone();
-            let plugin_name_ = plugin_name_.clone();
-            let task = RT.get().unwrap().spawn(async move {
-                PLUGIN_NAME
-                    .scope(plugin_name_, Bot::handler_drop(listen_clone))
-                    .await;
-            });
-            task_vec.push(task);
-        }
-
-        TASK_MANAGER.disable_plugin(&self.name);
-
-        self.enabled.send_modify(|v| {
-            *v = false;
-        });
-        self.listen.clear();
-        RT.get().unwrap().spawn(async move {
-            for task in task_vec {
-                let _ = task.await;
-            }
-        })
     }
 }
 
@@ -302,6 +81,7 @@ impl Bot {
     }
 
     /// 挂载插件的启动函数。
+    #[deprecated(since = "0.12.0", note = "请使用 `mount_plugin` 代替")]
     pub fn mount_main<T>(&mut self, name: T, version: T, main: Arc<KoviAsyncFn>)
     where
         String: From<T>,
@@ -309,7 +89,7 @@ impl Bot {
         let name = String::from(name);
         let version = String::from(version);
         let (tx, _rx) = watch::channel(true);
-        let bot_plugin = BotPlugin {
+        let bot_plugin = Plugin {
             enable_on_startup: true,
             enabled: tx,
             name: name.clone(),
@@ -325,6 +105,11 @@ impl Bot {
             access_list: AccessList::default(),
         };
         self.plugins.insert(name, bot_plugin);
+    }
+
+    /// 挂载插件的启动函数。
+    pub fn mount_plugin(&mut self, plugin: Plugin) {
+        self.plugins.insert(plugin.name.clone(), plugin);
     }
 
     /// 读取本地Kovi.conf.toml文件
@@ -583,6 +368,121 @@ impl Bot {
     }
 }
 
+#[derive(Debug, Deserialize, Serialize, Clone)]
+pub struct SendApi {
+    pub action: String,
+    pub params: Value,
+    pub echo: String,
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct ApiReturn {
+    pub status: String,
+    pub retcode: i32,
+    pub data: Value,
+    pub echo: String,
+}
+
+/// kovi的配置
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct KoviConf {
+    pub config: Config,
+    pub server: Server,
+}
+
+impl AsRef<KoviConf> for KoviConf {
+    fn as_ref(&self) -> &KoviConf {
+        self
+    }
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct Config {
+    pub main_admin: i64,
+    pub admins: Vec<i64>,
+    pub debug: bool,
+}
+
+impl KoviConf {
+    pub fn new(main_admin: i64, admins: Option<Vec<i64>>, server: Server, debug: bool) -> Self {
+        KoviConf {
+            config: Config {
+                main_admin,
+                admins: admins.unwrap_or_default(),
+                debug,
+            },
+            server,
+        }
+    }
+}
+
+/// bot信息结构体
+#[derive(Debug, Clone)]
+pub struct BotInformation {
+    pub main_admin: i64,
+    pub deputy_admins: HashSet<i64>,
+    pub server: Server,
+}
+/// server信息
+#[derive(Deserialize, Serialize, Debug, Clone)]
+pub struct Server {
+    pub host: Host,
+    pub port: u16,
+    pub access_token: String,
+    pub secure: bool,
+}
+impl Server {
+    pub fn new(host: Host, port: u16, access_token: String, secure: bool) -> Self {
+        Server {
+            host,
+            port,
+            access_token,
+            secure,
+        }
+    }
+}
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+#[serde(untagged)]
+pub enum Host {
+    IpAddr(IpAddr),
+    Domain(String),
+}
+impl Display for Host {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Host::IpAddr(ip) => write!(f, "{}", ip),
+            Host::Domain(domain) => write!(f, "{}", domain),
+        }
+    }
+}
+
+impl std::fmt::Display for ApiReturn {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "status: {}, retcode: {}, data: {}, echo: {}",
+            self.status, self.retcode, self.data, self.echo
+        )
+    }
+}
+
+impl std::fmt::Display for SendApi {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", serde_json::to_string(self).unwrap())
+    }
+}
+
+impl SendApi {
+    pub fn new(action: &str, params: Value, echo: &str) -> Self {
+        SendApi {
+            action: action.to_string(),
+            params,
+            echo: echo.to_string(),
+        }
+    }
+}
+
 /// 将配置文件写入磁盘
 fn config_file_write_and_return() -> Result<KoviConf, std::io::Error> {
     enum HostType {
@@ -743,9 +643,9 @@ macro_rules! build_bot {
             let mut bot = kovi::bot::Bot::build(&conf);
 
             $(
-                let (crate_name, crate_version) = $plugin::__kovi_get_plugin_info();
-                kovi::log::info!("Mounting plugin: {}", crate_name);
-                bot.mount_main(crate_name, crate_version, std::sync::Arc::new($plugin::__kovi_run_async_plugin));
+                let plugin = $plugin::__kovi_build_plugin();
+                kovi::log::info!("Mounting plugin: {}", &plugin.name);
+                bot.mount_plugin(plugin);
             )*
 
             bot.set_plugin_startup_use_file_ref();
